feat: 商品增加计价类型字段（number_type） 0 整数 1 小数；修改所有相关的字段类型为float64

+++ main/app/dto/req/instant.go
@@ -141,8 +141,8 @@ type InstantOrderSaleOrderMoveProductReq struct {
 
 // MoveProduct 移动商品
 type MoveProduct struct {
-	Uuid uint64 `json:"uuid"` // 销售订单商品UUID, 必填. 也能是顾客uuid、加钟uuid
-	Num  uint   `json:"num"`  // 移动数量, 必填
+	Uuid uint64  `json:"uuid"` // 销售订单商品UUID, 必填. 也能是顾客uuid、加钟uuid
+	Num  float64 `json:"num"`  // 移动数量, 必填
 }
 
 type InstantOrderMustPlanConfirmReq struct {
