+++ main/app/printer/template/statement_order_img.go
@@ -316,13 +316,13 @@ func (t *statementOrderImgTemplate) GetPrintContent(
 
 	// 赠品金额 / 商品数量
 	freeMoney := float64(0)
-	productNum := uint(0)
+	productNum := float64(0)
 	// 自助餐顾客类型
 	for _, orderBuffetCustomer := range saleOrder.SaleOrderBuffetCustomerTypes {
 		if orderBuffetCustomer.IsDelete() {
 			continue
 		}
-		productNum += orderBuffetCustomer.Num
+		productNum += float64(orderBuffetCustomer.Num)
 		buffetNameText := orderBuffetCustomer.BuffetPackage.MultiLanguageName.GetNameByLang(t.base.Lang)
 		if orderBuffetCustomer.BuffetCustomerTypePrice.BuffetCustomerType.Name != "" {
 			buffetNameText += "\n(" + orderBuffetCustomer.BuffetCustomerTypePrice.BuffetCustomerType.Name + ")"
@@ -339,7 +339,7 @@ func (t *statementOrderImgTemplate) GetPrintContent(
 		if delay.IsDelete() {
 			continue
 		}
-		productNum += delay.Num
+		productNum += float64(delay.Num)
 		originPrice := delay.GetAmount()
 		img.PrintInColumns(
 			pkg.ColumnConfig{Text: delay.Name, Width: 310, Align: pkg.AlignLeft},
